Remove component interface from contextprops
