+++ src/.eslintrc.js
@@ -53,9 +53,7 @@ module.exports = {
       'files': [
         './preact/base-element.js',
         './core/contextprops.js',
-        './context/component-hooks.js',
-        './context/component-install.js',
-        './context/component.js',
+        './context/subscriber.js',
         './context/node.js',
         './context/prop.js',
         './context/scheduler.js',
