+++ src/context/index.js
@@ -17,25 +17,7 @@
 import {ContextNode} from './node';
 
 export {contextProp} from './prop';
-export {withMetaData} from './component-meta';
-export {
-  mountComponent,
-  unmountComponent,
-  subscribe,
-  unsubscribe,
-  useMountComponent,
-  useUnmountComponent,
-  useSubscribe,
-  useUnsubscribe,
-} from './component-install';
-export {
-  useRef,
-  useMemo,
-  useDisposableMemo,
-  useSyncEffect,
-  useSetProp,
-  useRemoveProp,
-} from './component-hooks';
+export {subscribe, unsubscribe} from './subscriber';
 
 /**
  * Direct slot assignment. Works the same way as shadow slots, but does not
